refactor!: move common logger implementations to platform-api
