+++ platform-api/src/main/java/dev/hypera/chameleon/platform/logger/ChameleonJavaLogger.java
@@ -21,30 +21,35 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package dev.hypera.chameleon.logger;
+package dev.hypera.chameleon.platform.logger;
 
+import dev.hypera.chameleon.logger.AbstractChameleonLogger;
 import dev.hypera.chameleon.util.logger.FormattedMessage;
 import dev.hypera.chameleon.util.logger.MessageFormatter;
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.Logger;
-import org.jetbrains.annotations.ApiStatus.Internal;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
- * Log4J Chameleon logger implementation.
+ * Java Chameleon logger implementation.
  */
-@Internal
-public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
+public final class ChameleonJavaLogger extends AbstractChameleonLogger {
+
+    private static final @NotNull Level LEVEL_TRACE = Level.FINER;
+    private static final @NotNull Level LEVEL_DEBUG = Level.FINE;
+    private static final @NotNull Level LEVEL_INFO = Level.INFO;
+    private static final @NotNull Level LEVEL_WARN = Level.WARNING;
+    private static final @NotNull Level LEVEL_ERROR = Level.SEVERE;
 
     private final @NotNull Logger logger;
 
     /**
-     * Chameleon Log4J logger constructor.
+     * Chameleon Java logger constructor.
      *
-     * @param logger Log4J logger instance to use.
+     * @param logger Java logger instance to use.
      */
-    public ChameleonLog4jLogger(@NotNull Logger logger) {
+    public ChameleonJavaLogger(@NotNull Logger logger) {
         this.logger = logger;
     }
 
@@ -53,7 +58,9 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
      */
     @Override
     public void trace(@NotNull String msg) {
-        this.logger.trace(msg);
+        if (isTraceEnabled()) {
+            this.logger.log(LEVEL_TRACE, msg);
+        }
     }
 
     /**
@@ -62,7 +69,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void trace(@NotNull String format, @Nullable Object arg) {
         if (isTraceEnabled()) {
-            logFormatted(Level.TRACE, format, arg);
+            logFormatted(LEVEL_TRACE, format, arg);
         }
     }
 
@@ -72,7 +79,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void trace(@NotNull String format, @Nullable Object arg1, @Nullable Object arg2) {
         if (isTraceEnabled()) {
-            logFormatted(Level.TRACE, format, arg1, arg2);
+            logFormatted(LEVEL_TRACE, format, arg1, arg2);
         }
     }
 
@@ -82,7 +89,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void trace(@NotNull String format, @Nullable Object @NotNull ... arguments) {
         if (isTraceEnabled()) {
-            logFormatted(Level.TRACE, format, arguments);
+            logFormatted(LEVEL_TRACE, format, arguments);
         }
     }
 
@@ -92,7 +99,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void trace(@NotNull String msg, @Nullable Throwable t) {
         if (isTraceEnabled()) {
-            this.logger.trace(msg, t);
+            this.logger.log(LEVEL_TRACE, msg, t);
         }
     }
 
@@ -102,7 +109,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void debug(@NotNull String msg) {
         if (isDebugEnabled()) {
-            this.logger.debug(msg);
+            this.logger.log(LEVEL_DEBUG, msg);
         }
     }
 
@@ -112,7 +119,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void debug(@NotNull String format, @Nullable Object arg) {
         if (isDebugEnabled()) {
-            logFormatted(Level.DEBUG, format, arg);
+            logFormatted(LEVEL_DEBUG, format, arg);
         }
     }
 
@@ -122,7 +129,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void debug(@NotNull String format, @Nullable Object arg1, @Nullable Object arg2) {
         if (isDebugEnabled()) {
-            logFormatted(Level.DEBUG, format, arg1, arg2);
+            logFormatted(LEVEL_DEBUG, format, arg1, arg2);
         }
     }
 
@@ -132,7 +139,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void debug(@NotNull String format, @Nullable Object @NotNull ... arguments) {
         if (isDebugEnabled()) {
-            logFormatted(Level.DEBUG, format, arguments);
+            logFormatted(LEVEL_DEBUG, format, arguments);
         }
     }
 
@@ -142,7 +149,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void debug(@NotNull String msg, @Nullable Throwable t) {
         if (isDebugEnabled()) {
-            this.logger.debug(msg, t);
+            this.logger.log(LEVEL_DEBUG, msg, t);
         }
     }
 
@@ -152,7 +159,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void info(@NotNull String msg) {
         if (isInfoEnabled()) {
-            this.logger.info(msg);
+            this.logger.log(LEVEL_INFO, msg);
         }
     }
 
@@ -162,7 +169,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void info(@NotNull String format, @Nullable Object arg) {
         if (isInfoEnabled()) {
-            logFormatted(Level.INFO, format, arg);
+            logFormatted(LEVEL_INFO, format, arg);
         }
     }
 
@@ -172,7 +179,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void info(@NotNull String format, @Nullable Object arg1, @Nullable Object arg2) {
         if (isInfoEnabled()) {
-            logFormatted(Level.INFO, format, arg1, arg2);
+            logFormatted(LEVEL_INFO, format, arg1, arg2);
         }
     }
 
@@ -182,7 +189,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void info(@NotNull String format, @Nullable Object @NotNull ... arguments) {
         if (isInfoEnabled()) {
-            logFormatted(Level.INFO, format, arguments);
+            logFormatted(LEVEL_INFO, format, arguments);
         }
     }
 
@@ -192,7 +199,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void info(@NotNull String msg, @Nullable Throwable t) {
         if (isInfoEnabled()) {
-            this.logger.info(msg, t);
+            this.logger.log(LEVEL_INFO, msg, t);
         }
     }
 
@@ -202,7 +209,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void warn(@NotNull String msg) {
         if (isWarnEnabled()) {
-            this.logger.warn(msg);
+            this.logger.log(LEVEL_WARN, msg);
         }
     }
 
@@ -212,7 +219,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void warn(@NotNull String format, @Nullable Object arg) {
         if (isWarnEnabled()) {
-            logFormatted(Level.WARN, format, arg);
+            logFormatted(LEVEL_WARN, format, arg);
         }
     }
 
@@ -222,7 +229,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void warn(@NotNull String format, @Nullable Object arg1, @Nullable Object arg2) {
         if (isWarnEnabled()) {
-            logFormatted(Level.WARN, format, arg1, arg2);
+            logFormatted(LEVEL_WARN, format, arg1, arg2);
         }
     }
 
@@ -232,7 +239,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void warn(@NotNull String format, @Nullable Object @NotNull ... arguments) {
         if (isWarnEnabled()) {
-            logFormatted(Level.WARN, format, arguments);
+            logFormatted(LEVEL_WARN, format, arguments);
         }
     }
 
@@ -242,7 +249,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void warn(@NotNull String msg, @Nullable Throwable t) {
         if (isWarnEnabled()) {
-            this.logger.warn(msg, t);
+            this.logger.log(LEVEL_WARN, msg, t);
         }
     }
 
@@ -252,7 +259,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void error(@NotNull String msg) {
         if (isErrorEnabled()) {
-            this.logger.error(msg);
+            this.logger.log(LEVEL_ERROR, msg);
         }
     }
 
@@ -262,7 +269,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void error(@NotNull String format, @Nullable Object arg) {
         if (isErrorEnabled()) {
-            logFormatted(Level.ERROR, format, arg);
+            logFormatted(LEVEL_ERROR, format, arg);
         }
     }
 
@@ -272,7 +279,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void error(@NotNull String format, @Nullable Object arg1, @Nullable Object arg2) {
         if (isErrorEnabled()) {
-            logFormatted(Level.ERROR, format, arg1, arg2);
+            logFormatted(LEVEL_ERROR, format, arg1, arg2);
         }
     }
 
@@ -282,7 +289,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void error(@NotNull String format, @Nullable Object @NotNull ... arguments) {
         if (isErrorEnabled()) {
-            logFormatted(Level.ERROR, format, arguments);
+            logFormatted(LEVEL_ERROR, format, arguments);
         }
     }
 
@@ -292,7 +299,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
     @Override
     public void error(@NotNull String msg, @Nullable Throwable t) {
         if (isErrorEnabled()) {
-            this.logger.error(msg, t);
+            this.logger.log(LEVEL_ERROR, msg, t);
         }
     }
 
@@ -301,7 +308,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
      */
     @Override
     public boolean isTraceEnabled() {
-        return this.logger.isTraceEnabled();
+        return this.logger.isLoggable(LEVEL_TRACE);
     }
 
     /**
@@ -309,7 +316,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
      */
     @Override
     public boolean isDebugEnabled() {
-        return this.logger.isDebugEnabled();
+        return this.logger.isLoggable(LEVEL_DEBUG);
     }
 
     /**
@@ -317,7 +324,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
      */
     @Override
     public boolean isInfoEnabled() {
-        return this.logger.isInfoEnabled();
+        return this.logger.isLoggable(LEVEL_INFO);
     }
 
     /**
@@ -325,7 +332,7 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
      */
     @Override
     public boolean isWarnEnabled() {
-        return this.logger.isWarnEnabled();
+        return this.logger.isLoggable(LEVEL_WARN);
     }
 
     /**
@@ -333,16 +340,16 @@ public final class ChameleonLog4jLogger extends AbstractChameleonLogger {
      */
     @Override
     public boolean isErrorEnabled() {
-        return this.logger.isErrorEnabled();
+        return this.logger.isLoggable(LEVEL_ERROR);
     }
 
     private void logFormatted(@NotNull Level level, @NotNull String format, @Nullable Object @NotNull ... args) {
         FormattedMessage m = MessageFormatter.format(format, args);
         if (m.throwable() == null) {
-            this.logger.atLevel(level).log(m.message());
+            this.logger.log(level, m.message());
             return;
         }
-        this.logger.atLevel(level).withThrowable(m.throwable()).log(m.message());
+        this.logger.log(level, m.message(), m.throwable());
     }
 
 }
