+++ platform-api/src/main/java/dev/hypera/chameleon/platform/logger/ChameleonSlf4jLogger.java
@@ -21,8 +21,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package dev.hypera.chameleon.logger;
+package dev.hypera.chameleon.platform.logger;
 
+import dev.hypera.chameleon.logger.AbstractChameleonLogger;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
